fix(fsrepo.Peerstore): fix peerstore writing invalid json
