+++ p2p/node.go
@@ -206,56 +206,6 @@ func (n *QriNode) EncapsulatedAddresses() []ma.Multiaddr {
 	return res
 }
 
-// Peers returns a list of currently connected peer IDs
-func (n *QriNode) Peers() []peer.ID {
-	if n.Host == nil {
-		return []peer.ID{}
-	}
-	conns := n.Host.Network().Conns()
-	seen := make(map[peer.ID]struct{})
-	peers := make([]peer.ID, 0, len(conns))
-
-	for _, c := range conns {
-		p := c.LocalPeer()
-		if _, found := seen[p]; found {
-			continue
-		}
-
-		seen[p] = struct{}{}
-		peers = append(peers, p)
-	}
-
-	return peers
-}
-
-// AddQriPeer negotiates a connection with a peer to get their profile details
-// and peer list.
-func (n *QriNode) AddQriPeer(pinfo pstore.PeerInfo) error {
-	// add this peer to our store
-	n.QriPeers.AddAddrs(pinfo.ID, pinfo.Addrs, pstore.TempAddrTTL)
-
-	if _, err := n.RequestProfile(pinfo.ID); err != nil {
-		log.Debug(err.Error())
-		return err
-	}
-
-	return nil
-}
-
-// ConnectedPeers lists all IPFS connected peers
-func (n *QriNode) ConnectedPeers() []string {
-	if n.Host == nil {
-		return []string{}
-	}
-	conns := n.Host.Network().Conns()
-	peers := make([]string, len(conns))
-	for i, c := range conns {
-		peers[i] = c.RemotePeer().Pretty()
-	}
-
-	return peers
-}
-
 // ConnectedQriProfiles lists all connected peers that support the qri protocol
 func (n *QriNode) ConnectedQriProfiles() map[profile.ID]*profile.Profile {
 	if n.Host == nil {
@@ -270,32 +220,6 @@ func (n *QriNode) ConnectedQriProfiles() map[profile.ID]*profile.Profile {
 		}
 	}
 	return peers
-}
-
-// ConnectToPeer takes a raw peer ID & tries to work out a route to that
-// peer, explicitly connecting to them.
-func (n *QriNode) ConnectToPeer(pid peer.ID) error {
-	// first check for local peer info
-	if pinfo := n.Host.Peerstore().PeerInfo(pid); pinfo.ID.String() != "" {
-		_, err := n.RequestProfile(pinfo.ID)
-		return err
-	}
-
-	// attempt to use ipfs routing table to discover peer
-	ipfsnode, err := n.IPFSNode()
-	if err != nil {
-		log.Debug(err.Error())
-		return err
-	}
-
-	pinfo, err := ipfsnode.Routing.FindPeer(context.Background(), pid)
-	if err != nil {
-		log.Debug(err.Error())
-		return err
-	}
-
-	_, err = n.RequestProfile(pinfo.ID)
-	return err
 }
 
 // Context returns this node's context
