+++ p2p/peers.go
@@ -1,8 +1,11 @@
 package p2p
 
 import (
+	"context"
+
 	"github.com/qri-io/qri/repo/profile"
 
+	pstore "gx/ipfs/QmXauCuJzmzapetmC6W4TuDJLL1yFFrVzSHoWv8YdbmnxH/go-libp2p-peerstore"
 	peer "gx/ipfs/QmZoWKhxUmZ2seW4BzX6fJkNR8hh9PsGModr7q171yq2SS/go-libp2p-peer"
 )
 
@@ -34,4 +37,80 @@ func (n *QriNode) ClosestConnectedPeers(id profile.ID, max int) (pid []peer.ID) 
 	}
 
 	return
+}
+
+// AddQriPeer negotiates a connection with a peer to get their profile details
+// and peer list.
+func (n *QriNode) AddQriPeer(pinfo pstore.PeerInfo) error {
+	// add this peer to our store
+	n.QriPeers.AddAddrs(pinfo.ID, pinfo.Addrs, pstore.TempAddrTTL)
+
+	if _, err := n.RequestProfile(pinfo.ID); err != nil {
+		log.Debug(err.Error())
+		return err
+	}
+
+	return nil
+}
+
+// Peers returns a list of currently connected peer IDs
+func (n *QriNode) Peers() []peer.ID {
+	if n.Host == nil {
+		return []peer.ID{}
+	}
+	conns := n.Host.Network().Conns()
+	seen := make(map[peer.ID]struct{})
+	peers := make([]peer.ID, 0, len(conns))
+
+	for _, c := range conns {
+		p := c.LocalPeer()
+		if _, found := seen[p]; found {
+			continue
+		}
+
+		seen[p] = struct{}{}
+		peers = append(peers, p)
+	}
+
+	return peers
+}
+
+// ConnectedPeers lists all IPFS connected peers
+func (n *QriNode) ConnectedPeers() []string {
+	if n.Host == nil {
+		return []string{}
+	}
+	conns := n.Host.Network().Conns()
+	peers := make([]string, len(conns))
+	for i, c := range conns {
+		peers[i] = c.RemotePeer().Pretty()
+	}
+
+	return peers
+}
+
+// ConnectToPeer takes a raw peer ID & tries to work out a route to that
+// peer, explicitly connecting to them.
+func (n *QriNode) ConnectToPeer(pid peer.ID) error {
+	// first check for local peer info
+	if pinfo := n.Host.Peerstore().PeerInfo(pid); pinfo.ID.String() != "" {
+		_, err := n.RequestProfile(pinfo.ID)
+		return err
+	}
+
+	// attempt to use ipfs routing table to discover peer
+	ipfsnode, err := n.IPFSNode()
+	if err != nil {
+		log.Debug(err.Error())
+		return err
+	}
+
+	pinfo, err := ipfsnode.Routing.FindPeer(context.Background(), pid)
+	if err != nil {
+		log.Debug(err.Error())
+		return err
+	}
+
+	_, err = n.RequestProfile(pinfo.ID)
+	return err
 }
